+++ packages/editor/js/extensions/libs/typography/test/text-orientation.e2e.cy.js
@@ -21,9 +21,7 @@ describe('Text Orientation → Functionality', () => {
 		//
 		// Style 1
 		//
-		cy.getByAriaLabel(
-			'Text will display vertically from left to right with a mixed orientation'
-		).click();
+		cy.getByAriaLabel('Vertical LR Mixed').click();
 
 		//Check block
 		cy.getBlock('core/paragraph')
@@ -40,9 +38,7 @@ describe('Text Orientation → Functionality', () => {
 		//
 		// Style 2
 		//
-		cy.getByAriaLabel(
-			'Text will display vertically from right to left with a mixed orientation'
-		).click();
+		cy.getByAriaLabel('Vertical RL Mixed').click();
 
 		//Check block
 		cy.getBlock('core/paragraph')
@@ -59,9 +55,7 @@ describe('Text Orientation → Functionality', () => {
 		//
 		// Style 3
 		//
-		cy.getByAriaLabel(
-			'Text will appear vertically from left to right with an upright orientation'
-		).click();
+		cy.getByAriaLabel('Vertical LR Upright').click();
 
 		//Check block
 		cy.getBlock('core/paragraph')
@@ -78,9 +72,7 @@ describe('Text Orientation → Functionality', () => {
 		//
 		// Style 4
 		//
-		cy.getByAriaLabel(
-			'Text will appear vertically from right to left with an upright orientation'
-		).click();
+		cy.getByAriaLabel('Vertical RL Upright').click();
 
 		//Check block
 		cy.getBlock('core/paragraph')
@@ -112,9 +104,7 @@ describe('Text Orientation → Functionality', () => {
 		});
 
 		// Switch to style 1
-		cy.getByAriaLabel(
-			'Text will display vertically from left to right with a mixed orientation'
-		).click();
+		cy.getByAriaLabel('Vertical LR Mixed').click();
 
 		//Check frontend
 		savePage();